Death animation for enemies!
At the moment the enemy has valid hitbox during death animation, can be fixed if we modify the if condition in the collision. The variable is self.alive so we just have to add that there. (I didn't want to tamper too much with your part of code to avoid unnecessary  conflicts)